refact: factor out useSearchRouterAndHandler effect to file
